requirements.txt
fastapi~=0.116.1
uvicorn
langchain~=0.3.26
langchain-community~=0.3.27
langchain-core~=0.3.68
tiktoken~=0.9.0
sentence-transformers
psycopg2-binary~=2.9.10
pydantic~=2.11.7
python-multipart
transformers~=4.53.0